TSD-320: Github issues 85; refactor add to cart analytics to match conversion; fixed conversion sku; moved preparePayload to service
