+++ Api/AnalyticsServiceInterface.php
@@ -7,14 +7,17 @@ declare(strict_types=1);
 
 namespace Pixlee\Pixlee\Api;
 
+use Magento\Store\Api\Data\StoreInterface;
+
 interface AnalyticsServiceInterface
 {
     /**
-     * @param $event
-     * @param $payload
+     * @param String $event
+     * @param array $payload
+     * @param StoreInterface $store
      * @return mixed
      */
-    public function sendEvent($event, $payload);
+    public function sendEvent($event, $payload, $store);
 
     /**
      * @param $path
