+++ Service/Analytics.php
@@ -7,10 +7,15 @@ declare(strict_types=1);
 
 namespace Pixlee\Pixlee\Service;
 
+use Magento\Framework\App\ProductMetadataInterface;
 use Magento\Framework\HTTP\Client\Curl;
 use Magento\Framework\Serialize\SerializerInterface;
+use Magento\Store\Model\ScopeInterface;
 use Pixlee\Pixlee\Api\AnalyticsServiceInterface;
+use Pixlee\Pixlee\Model\Config\Api;
+use Pixlee\Pixlee\Model\CookieManager;
 use Pixlee\Pixlee\Model\Logger\PixleeLogger;
+use Pixlee\Pixlee\Model\Pixlee;
 
 class Analytics implements AnalyticsServiceInterface
 {
@@ -27,34 +32,61 @@ class Analytics implements AnalyticsServiceInterface
      * @var SerializerInterface
      */
     protected $serializer;
+    /**
+     * @var Api
+     */
+    protected $apiConfig;
+    /**
+     * @var CookieManager
+     */
+    protected $cookieManager;
+    /**
+     * @var Pixlee
+     */
+    protected $pixlee;
+    /**
+     * @var ProductMetadataInterface
+     */
+    protected $productMetadata;
 
     /**
+     * @param Api $apiConfig
+     * @param CookieManager $cookieManager
      * @param Curl $curl
-     * @param SerializerInterface $serializer
      * @param PixleeLogger $logger
+     * @param Pixlee $pixlee
+     * @param ProductMetadataInterface $productMetadata
+     * @param SerializerInterface $serializer
      */
     public function __construct(
+        Api $apiConfig,
+        CookieManager $cookieManager,
         Curl $curl,
-        SerializerInterface $serializer,
-        PixleeLogger $logger
+        PixleeLogger $logger,
+        Pixlee $pixlee,
+        ProductMetadataInterface $productMetadata,
+        SerializerInterface $serializer
     ) {
+        $this->apiConfig = $apiConfig;
+        $this->cookieManager = $cookieManager;
         $this->curl = $curl;
-        $this->serializer = $serializer;
         $this->logger = $logger;
+        $this->pixlee = $pixlee;
+        $this->productMetadata = $productMetadata;
+        $this->serializer = $serializer;
     }
 
     /**
      * @inheritdoc
      */
-    public function sendEvent($event, $payload)
+    public function sendEvent($event, $payload, $store)
     {
         $urls = [
             'addToCart' => 'addToCart',
-            'checkoutStart' => 'checkoutStart',
             'checkoutSuccess' => 'conversion',
-            'removeFromCart' => 'removeFromCart'
         ];
         if ($payload && isset($urls[$event])) {
+            $payload = $this->preparePayload($store, $payload);
             $path = "events/{$urls[$event]}";
             $response = $this->post($path, $payload);
 
@@ -65,6 +97,50 @@ class Analytics implements AnalyticsServiceInterface
         }
 
         $this->logger->addInfo("Pixlee Analytics: Event not sent - " . $this->serializer->serialize($payload));
+        return false;
+    }
+
+    /**
+     * @param $store
+     * @param $extraData
+     * @return false|string
+     */
+    public function preparePayload($store, $extraData = [])
+    {
+        if ($payload = $this->getPixleeCookie()) {
+            foreach ($extraData as $key => $value) {
+                // Don't overwrite existing data.
+                if (!isset($payload[$key])) {
+                    $payload[$key] = $value;
+                }
+            }
+            // Required key/value pairs not in the payload by default.
+            $payload['API_KEY']= $this->apiConfig->getPrivateApiKey(ScopeInterface::SCOPE_STORES, $store->getCode());
+            $payload['distinct_user_hash'] = $payload['CURRENT_PIXLEE_USER_ID'];
+            $payload['ecommerce_platform'] = Pixlee::PLATFORM;
+            $payload['ecommerce_platform_version'] = $this->productMetadata->getVersion();
+            $payload['version_hash'] = $this->pixlee->getExtensionVersion();
+            $payload['region_code'] = $store->getCode();
+            $serializedPayload = $this->serializer->serialize($payload);
+
+            $this->logger->addInfo("Sending payload: " . $serializedPayload);
+            return $serializedPayload;
+        }
+
+        $this->logger->addInfo("Analytics event not sent because the cookie wasn't found");
+        return false;
+    }
+
+    /**
+     * @return array|bool
+     */
+    public function getPixleeCookie()
+    {
+        $cookie = $this->cookieManager->get();
+        if (isset($cookie)) {
+            return $this->serializer->unserialize($cookie);
+        }
+
         return false;
     }
 
@@ -98,7 +174,7 @@ class Analytics implements AnalyticsServiceInterface
      * @param $responseCode
      * @return bool
      */
-    protected function isValidResponse($response, $responseCode)
+    public function isValidResponse($response, $responseCode)
     {
         if (!$this->isBetween($responseCode, 200, 299)) {
             $this->logger->addInfo("Pixlee Analytics: HTTP $responseCode response");
@@ -123,7 +199,7 @@ class Analytics implements AnalyticsServiceInterface
      * @param $high
      * @return bool
      */
-    protected function isBetween($theNum, $low, $high)
+    public function isBetween($theNum, $low, $high)
     {
         if ($theNum >= $low && $theNum <= $high) {
             return true;
